chore: Fixed loading onnx model
